harmofl-init/requirements.txt
Cython==3.0.0
numpy==1.25.0
opencv_python==4.8.0.74
Pillow==10.0.0
setuptools==68.0.0
SimpleITK==2.2.1
torch==2.0.1
torchvision==0.15.2
